adds local verification method for biscuit tokens
This commit adds the ability to locally verify a biscuit token if there
is a public_key configured. If the local verification fails for whatever
reason, or there is no public_key configured, it will fall back to
making a network request for verification.

Also added in this change are some slightly better HessraConfig
ergonomics by adding a builder and a way to create a HessraClient from a
HessraConfig.