removes hessra-macros
The API is in a state of flux right now. I am removing the macros crate
to reduce the amount of work that is needed for this SDK as we iterate,
add, and remove things from the authorization API.

Once the API has been stabilized, we will revisit the macros and add
them then.