hessra-sdk: updates SDK to have local/remote abstraction
This adds specific local/remote calls to verify tokens and a new call
that will try and first verify a token locally before attempting a
remote verification. This way, consumers of the SDK can choose what they
would like, or have the SDK choose for them based on what is available.