fixing up issues
trying to fix up some issues with the http3 client. it's still broken in
my testing and it looks like most of the problems are around URL/Address
handling in the h3/quinn crates.

I'm going to check in these changes for now, but shelve work on the
http3 feature for now and just wait until reqwest's http3 support is
ready then just use that. Or, if somebody wants http3 sooner I'll iron
out what's broken.