replaces http3 client with reqwest
this replaces the underlying http3 client (h3, h3-quinn) with reqwest.
To use the http3 client, the cargo feature flag "http3" is needed along
with RUSTFLAGS="--cfg reqwest_unstable".

ex: RUSTFLAGS="--cfg reqwest_unstable" cargo build --features http3

Anything using the SDK that wishes to use the HTTP3 client will need to
provide the environment variable as well.

stable http3 support for reqwest looks like it is being worked on right
now, so it shouldn't be long that the env variable is necessary.

https://github.com/seanmonstar/reqwest/issues/2303