hessra-token: adds more testing
This adds more testing by using some generated long-lived test tokens by
the authorization service and putting them in test_tokens.json.

In doing this testing I was able to find an issue with the authority
block of the token incorrectly setting the subject, when just the
authorization block should be adding the subject fact.

That issue was fixed in the authorization service.

Something to note on service_chain tokens currently: the authorizer or
SDK user needs to know if it is a node in a service chain and if it
should be calling verify_token or verify_service_chain_token. Calling
verify_token on a service_chain token issued by the authorization
service that hasn't visited any nodes in the service chain will verify.
This is on purpose so that the first node in the service chain can
verify the token before adding its service node block to it.

In the future, we may add some kind of token type system to the
authority block that will help the SDK automatically determine how to
best authorize a given token without as much knowledge needed by the
user.