hessra-pgrx: update pgrx to 14.1 and remove from workspace
I've updated pgrx to 14.1 again since we are no longer using pg-trunk or
tembo and don't need to rely on the build containers that were missing.

I'm also removing the pgrx container from the workspace so it doesn't
build when the rest of the workspace is built. Running "cargo build" in
the workspace with the pgrx crate as a member causes a build failure for
pgrx and missing symbols on MacOS, but building just the pgrx crate
works fine.