Adds ability to set operation for tokens
This change adds the ability to formally set the operation for which a
subject is allowed for a resource on token creation and verification.

Previously, there was a hard-coded read+write operation added. This
changes the API to add a single allowed operation for a given token that
can be set to any string.

Tests, example code, and dependent crates should all be updated. I will
be updating the authorization service API and typescript API client as
well.