hessra-pgrx: updates README
This fixes some errors in the README, removing a function that wasn't
actually added and adding "operation" to the args of functions.

Some near future work: the component_name arg should be removed from
verification function calls. Since postgres should be the end of the
line for verifying a token, it doesn't make sense that you'd want to
only verify up to a certain point in a service chain.