fixes service chain nomenclature
I started off calling the service chain feature "attenuated" tokens. But
this isn't true. Attenuation is the act of restricting. So for an
authorization token that means restricting the scope of its permissions.

The service chain feature is actually more of an attestation. The nodes
in the service chain are adding context that builds the complete
provenance of the token and its path through a system. So it is not
really changing the scope of the token.

The service chain feature, for a given node, is literally saying I have
verified this token and am putting my signature on it saying that I have
done so, which is attestation.