hessra-api: updates documentation, README, and bumps version
There were some errors, discrepencies, and lack of clarity in the code
documentation and README. I took a pass to try and fix those issues and
make things more clear.

I've also bumped the crate to v0.3.0. This will likely be the next
version published to crates.io and tagged in github.