hessra-sdk/api: adds identity_token
This adds full use of identity tokens between the API/SDK and the
service. Identity tokens can be minted by the service for a particular
identity if that identity requests one and it is configured to get one.
The identity requesting an identity token must authenticate to the
service some other way (at this point, just mTLS).

Then, the identity token can be used to authenticate the identity or any
of its delegates to get authorization tokens based on configured policy
in the authorization service.

The biggest change is in use of the API/SDK. The only endpoint that
requires mTLS now is the one to request the initial identity token. This
means the HessraClient and Hessra SDK objects can now be created without
an mTLS client cert/key pair.

Also, the service functions in this way for mTLS and/or identity tokens:
mTLS alone: identity in certificate is used for authentication
mTLS + identity token: identities in cert and token must match
identity token alone: can be used for full authentication