hessra-cli: adds public key caching and integration test
this adds public key caching to the hessra CLI so that any subsequent
operations using the same authorization namespace don't require an API
call to the service to get the public key.

This also adds an example program that invokes the CLI and adds it to
the integration testing github workflow.