hessra-cli: tweaks delegation UX
This changes the delegation UX for the CLI. First, the token will always
be printed to stdout now. In addition, you can pass the --token-only
argument and have just the token be printed so that you can pipe/insert
more easily into other commands. Second, delegated token identities will
only be saved to ~/.hessra/ if given the --save-as command. Otherwise,
it will just be outputted to stdout.