hessra-token-identity: add bearer token support
This adds another verifier that can be used to treat an identity token
as a bearer token. What this means in practice, is that the token's
signature and expiration will be checked, but the actual subject encoded
in the token will not be. This is useful for using a hessra identity
token to replace a static API key, refresh token, or other similar
rudimentary identity with something that's better with built in
expiration.

Using a token like this can be a drop-in replacement for a rudimentary
API key implementation that gets you improvements right away (token
expiration and no need to track token state yourself) while giving a
clear upgrade path to better security.