hessra-token-identity: adds inspect methods
this adds methods to help inspect identity tokens and extract useful
information. When using a hessra identity token in place of a
rudimentary bearer token, you can use these methods to determine who the
actor/subject encoded in the token is and get useful information about
expiration for logs or displaying to users or tracking.

also, you can use it to get the most specific delegated identity from a
delegated identity token.