hessra-cli: adds symlink to certs folder in workspace
The hessra-cli defaults to use the CA in the certs folder if one isn't
provided. This seems like an okay thing to do for now since it would
help test the cli against the deployed test.hessra.net instance.

It should be noted however, that people using the CLI should pin the CA
cert themselves instead of relying on the one baked into the repository.

This is a bit ugly for the SDK's UX, especially around local delegation
and verification. I'll smooth it out in the future.