hessra-cli: improves ux
This improves the ux of using the hessra cli a bit. I added a ca_cert
public API endpoint to the service and hessra-api. The hessra cli now
makes use of this to automatically get the cert chain for verifying the
server with future commands.

Also in this change is organizing this configuration (cert chain, public
key) and tokens by server, in case you use multiple (dev/prod). In
addition, you can set a default server for easier operations without
providing the arg each time.

The goal is to get to the minimal set of args a user needs to provide
when using the hessra cli.