hessra-cli: changes "hessra init" to be interactive
The command "hessra init" is now interactive to help set up an auth
server for the first time or add a new one. It will summarize existing
configured auth servers if there are any, and then asks for the name of
the new auth server. Finally, it will ask the user if they want to set
it as the default.

The command "hessra config init" is left as-is and is the programmatic
interface for setting configuration.