hessra-token-identity: adds domain restricted, builder pattern
This adds the ability to create an identity token that is domain
restricted (think audience or application restricted). This is done by
adding a "check if domain(foo)" to the authority block of the identity
token.

This change also implements a builder pattern for identity tokens to
better help with building identity tokens that may or may not be
delegatable and may or may not be domain restricted.