hessra-token-authz: move capability verification
This change moves the verification of the capability

right(subject, resource, operation) to the authority block as a check
instead of being in the authorization block as an authorizer. This will
make it impossible for verifiers to not provide a correct subject,
operation, and resource.

This is also necessary to introduce a new token type in an upcoming
change.

This should effectively be a no-op. New tokens with the check should
also verify cleanly with the old authorizer. Although old tokens won't
verify with the new authorizer.