hessra-token-authz: adds capability_verification
This adds the ability to verify as a pure capability token, ignoring the
identity/subject and only authorizing based on the actual capability:
resource + operation.

This is useful and preferred for things like microservices that don't
care about identity and only care that a given request has the right
permissions to do what it is asking (like write telemetry).