hessra-token-identity: adds mechanic to associate domain and subject
This adds a rule to domain restricted identity tokens that lets a
verifier be sure that the subject minted for that domain is directly
associated with it by the minter using only the token.

Without this rule, it is feasible for ambiguity during verification
since checks can be added by third parties to other identity token
types.