hessra-cli: save CA provided by command line args
when providing a CA via command line arg, it isn't saved. this means
that one CLI invocation, which may have multiple API calls, might try
and use a previously saved CA or the system CAs. this worked for
deployed services with properly signed CAs but failed for local testing
with self-signed CAs.

this change should make the UX of the CLI correct in all situations